EH-424: Move organisation select to own component
