+++ shared/components/OrganisationDropdown.md
@@ -1,0 +1,10 @@
+### OrganisationDropdown with default values
+
+```js
+const { OrganisationDropdown } = require("components/OrganisationDropdown")
+;<OrganisationDropdown
+   oids={["123", "234"]}
+   onChange={console.log}
+   value={"123"}
+ />
+```
